requirements.txt
PyYAML
jsonschema
tensorflow
matplotlib
termcolor
numpy
scipy
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
